@@ -1,5 +1,5 @@
-//go:build darwin
-// +build darwin
+//go:build linux
+// +build linux
 
 package rime
 
@@ -9,11 +9,10 @@ import (
 	"path/filepath"
 )
 
-// 获取 macOS Rime 配置目录
 func getRimeDirForPlatform() string {
 	u, err := user.Current()
 	if err != nil {
 		log.Fatalln(err)
 	}
-	return filepath.Join(u.HomeDir, "Library/Rime")
+	return filepath.Join(u.HomeDir, ".config", "rime")
 }
