@@ -56,7 +56,6 @@ var polyphones = map[string]string{
 	"MAC地址 > 地":     "di",
 	"OK了 > 了":       "le",
 	"OK了吗 > 了":      "le",
-	"圈X > 圈":        "quan",
 	"A型血 > 血":       "xue",
 	"A血型 > 血":       "xue",
 	"B型血 > 血":       "xue",
@@ -92,6 +91,10 @@ var polyphones = map[string]string{
 	"Chrome系 > 系":   "xi",
 	"QQ游戏大厅 > 大":    "da",
 	"QQ飞车 > 车":      "che",
+	"2G网络 > 络":      "luo",
+	"3G网络 > 络":      "luo",
+	"4G网络 > 络":      "luo",
+	"5G网络 > 络":      "luo",
 }
 
 var digitMap = map[string]string{
@@ -118,330 +121,341 @@ type schema struct {
 	file              *os.File
 }
 
-var doublePinyin = schema{
-	name:            "cn_en_double_pinyin",
-	desc:            "自然码双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_double_pinyin.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "aa",
-		"-e-":   "ee",
-		"-o-":   "oo",
-		"-ai-":  "ai",
-		"-ei-":  "ei",
-		"-ou-":  "ou",
-		"-an-":  "an",
-		"-en-":  "en",
-		"-ang-": "ah",
-		"-eng-": "eg",
-		"-ao-":  "ao",
-		"-er-":  "er",
-		// zh ch sh
-		"zh": "v",
-		"ch": "i",
-		"sh": "u",
-		// 韵母
-		"iu":   "q",
-		"ia":   "w",
-		"ua":   "w",
-		"uan":  "r",
-		"ue":   "t",
-		"ve":   "t",
-		"ing":  "y",
-		"uai":  "y",
-		"uo":   "o",
-		"un":   "p",
-		"iong": "s",
-		"ong":  "s",
-		"iang": "d",
-		"uang": "d",
-		"en":   "f",
-		"eng":  "g",
-		"ang":  "h",
-		"an":   "j",
-		"ao":   "k",
-		"ai":   "l",
-		"ei":   "z",
-		"ie":   "x",
-		"iao":  "c",
-		"ui":   "v",
-		"ou":   "b",
-		"in":   "n",
-		"ian":  "m",
-	},
-}
+var (
+	doublePinyin        schema
+	doublePinyinFlypy   schema
+	doublePinyinMSPY    schema
+	doublePinyinSogou   schema
+	doublePinyinZiGuang schema
+	doublePinyinABC     schema
+)
 
-var doublePinyinFlypy = schema{
-	name:            "cn_en_flypy",
-	desc:            "小鹤双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_flypy.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "aa",
-		"-e-":   "ee",
-		"-o-":   "oo",
-		"-ai-":  "ai",
-		"-ei-":  "ei",
-		"-ou-":  "ou",
-		"-an-":  "an",
-		"-en-":  "en",
-		"-ang-": "ah",
-		"-eng-": "eg",
-		"-ao-":  "ao",
-		"-er-":  "er",
-		// zh ch sh
-		"zh": "v",
-		"ch": "i",
-		"sh": "u",
-		// 韵母
-		"iu":   "q",
-		"ei":   "w",
-		"uan":  "r",
-		"ue":   "t",
-		"ve":   "t",
-		"un":   "y",
-		"uo":   "o",
-		"ie":   "p",
-		"iong": "s",
-		"ong":  "s",
-		"ai":   "d",
-		"en":   "f",
-		"eng":  "g",
-		"ang":  "h",
-		"an":   "j",
-		"ing":  "k",
-		"uai":  "k",
-		"iang": "l",
-		"uang": "l",
-		"ou":   "z",
-		"ia":   "x",
-		"ua":   "x",
-		"ao":   "c",
-		"ui":   "v",
-		"in":   "b",
-		"iao":  "n",
-		"ian":  "m",
-	},
-}
+func initSchemas() {
+	doublePinyin = schema{
+		name:            "cn_en_double_pinyin",
+		desc:            "自然码双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_double_pinyin.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "aa",
+			"-e-":   "ee",
+			"-o-":   "oo",
+			"-ai-":  "ai",
+			"-ei-":  "ei",
+			"-ou-":  "ou",
+			"-an-":  "an",
+			"-en-":  "en",
+			"-ang-": "ah",
+			"-eng-": "eg",
+			"-ao-":  "ao",
+			"-er-":  "er",
+			// zh ch sh
+			"zh": "v",
+			"ch": "i",
+			"sh": "u",
+			// 韵母
+			"iu":   "q",
+			"ia":   "w",
+			"ua":   "w",
+			"uan":  "r",
+			"ue":   "t",
+			"ve":   "t",
+			"ing":  "y",
+			"uai":  "y",
+			"uo":   "o",
+			"un":   "p",
+			"iong": "s",
+			"ong":  "s",
+			"iang": "d",
+			"uang": "d",
+			"en":   "f",
+			"eng":  "g",
+			"ang":  "h",
+			"an":   "j",
+			"ao":   "k",
+			"ai":   "l",
+			"ei":   "z",
+			"ie":   "x",
+			"iao":  "c",
+			"ui":   "v",
+			"ou":   "b",
+			"in":   "n",
+			"ian":  "m",
+		},
+	}
 
-var doublePinyinMSPY = schema{
-	name:            "cn_en_mspy",
-	desc:            "微软双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_mspy.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "oa",
-		"-e-":   "oe",
-		"-o-":   "oo",
-		"-ai-":  "ol",
-		"-ei-":  "oz",
-		"-ou-":  "ob",
-		"-an-":  "oj",
-		"-en-":  "of",
-		"-ang-": "oh",
-		"-eng-": "og",
-		"-ao-":  "ok",
-		"-er-":  "or",
-		// zh ch sh
-		"zh": "v",
-		"ch": "i",
-		"sh": "u",
-		// 韵母
-		"iu":   "q",
-		"ia":   "w",
-		"ua":   "w",
-		"er":   "r",
-		"uan":  "r",
-		"ue":   "t",
-		"uai":  "y",
-		"uo":   "o",
-		"un":   "p",
-		"iong": "s",
-		"ong":  "s",
-		"iang": "d",
-		"uang": "d",
-		"en":   "f",
-		"eng":  "g",
-		"ang":  "h",
-		"an":   "j",
-		"ao":   "k",
-		"ai":   "l",
-		"ing":  ";",
-		"ei":   "z",
-		"ie":   "x",
-		"iao":  "c",
-		"ui":   "v",
-		"ve":   "v",
-		"ou":   "b",
-		"in":   "n",
-		"ian":  "m",
-	},
-}
+	doublePinyinFlypy = schema{
+		name:            "cn_en_flypy",
+		desc:            "小鹤双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_flypy.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "aa",
+			"-e-":   "ee",
+			"-o-":   "oo",
+			"-ai-":  "ai",
+			"-ei-":  "ei",
+			"-ou-":  "ou",
+			"-an-":  "an",
+			"-en-":  "en",
+			"-ang-": "ah",
+			"-eng-": "eg",
+			"-ao-":  "ao",
+			"-er-":  "er",
+			// zh ch sh
+			"zh": "v",
+			"ch": "i",
+			"sh": "u",
+			// 韵母
+			"iu":   "q",
+			"ei":   "w",
+			"uan":  "r",
+			"ue":   "t",
+			"ve":   "t",
+			"un":   "y",
+			"uo":   "o",
+			"ie":   "p",
+			"iong": "s",
+			"ong":  "s",
+			"ai":   "d",
+			"en":   "f",
+			"eng":  "g",
+			"ang":  "h",
+			"an":   "j",
+			"ing":  "k",
+			"uai":  "k",
+			"iang": "l",
+			"uang": "l",
+			"ou":   "z",
+			"ia":   "x",
+			"ua":   "x",
+			"ao":   "c",
+			"ui":   "v",
+			"in":   "b",
+			"iao":  "n",
+			"ian":  "m",
+		},
+	}
 
-var doublePinyinSogou = schema{
-	name:            "cn_en_sogou",
-	desc:            "搜狗双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_sogou.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "oa",
-		"-e-":   "oe",
-		"-o-":   "oo",
-		"-ai-":  "ol",
-		"-ei-":  "oz",
-		"-ou-":  "ob",
-		"-an-":  "oj",
-		"-en-":  "of",
-		"-ang-": "oh",
-		"-eng-": "og",
-		"-ao-":  "ok",
-		"-er-":  "or",
-		// zh ch sh
-		"zh": "v",
-		"ch": "i",
-		"sh": "u",
-		// 韵母
-		"iu":   "q",
-		"ia":   "w",
-		"ua":   "w",
-		"er":   "r",
-		"uan":  "r",
-		"ue":   "t",
-		"uai":  "y",
-		"uo":   "o",
-		"un":   "p",
-		"iong": "s",
-		"ong":  "s",
-		"iang": "d",
-		"uang": "d",
-		"en":   "f",
-		"eng":  "g",
-		"ang":  "h",
-		"an":   "j",
-		"ao":   "k",
-		"ai":   "l",
-		"ing":  ";",
-		"ei":   "z",
-		"ie":   "x",
-		"iao":  "c",
-		"ui":   "v",
-		"ve":   "t",
-		"ou":   "b",
-		"in":   "n",
-		"ian":  "m",
-	},
-}
+	doublePinyinMSPY = schema{
+		name:            "cn_en_mspy",
+		desc:            "微软双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_mspy.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "oa",
+			"-e-":   "oe",
+			"-o-":   "oo",
+			"-ai-":  "ol",
+			"-ei-":  "oz",
+			"-ou-":  "ob",
+			"-an-":  "oj",
+			"-en-":  "of",
+			"-ang-": "oh",
+			"-eng-": "og",
+			"-ao-":  "ok",
+			"-er-":  "or",
+			// zh ch sh
+			"zh": "v",
+			"ch": "i",
+			"sh": "u",
+			// 韵母
+			"iu":   "q",
+			"ia":   "w",
+			"ua":   "w",
+			"er":   "r",
+			"uan":  "r",
+			"ue":   "t",
+			"uai":  "y",
+			"uo":   "o",
+			"un":   "p",
+			"iong": "s",
+			"ong":  "s",
+			"iang": "d",
+			"uang": "d",
+			"en":   "f",
+			"eng":  "g",
+			"ang":  "h",
+			"an":   "j",
+			"ao":   "k",
+			"ai":   "l",
+			"ing":  ";",
+			"ei":   "z",
+			"ie":   "x",
+			"iao":  "c",
+			"ui":   "v",
+			"ve":   "v",
+			"ou":   "b",
+			"in":   "n",
+			"ian":  "m",
+		},
+	}
 
-var doublePinyinZiGuang = schema{
-	name:            "cn_en_ziguang",
-	desc:            "紫光双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_ziguang.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "oa",
-		"-e-":   "oe",
-		"-o-":   "oo",
-		"-ai-":  "op",
-		"-ei-":  "ok",
-		"-ou-":  "oz",
-		"-an-":  "or",
-		"-en-":  "ow",
-		"-ang-": "os",
-		"-eng-": "ot",
-		"-ao-":  "oq",
-		"-er-":  "oj",
-		// zh ch sh
-		"zh": "u",
-		"ch": "a",
-		"sh": "i",
-		// 韵母
-		"ao":   "q",
-		"en":   "w",
-		"an":   "r",
-		"eng":  "t",
-		"in":   "y",
-		"uai":  "y",
-		"uo":   "o",
-		"ai":   "p",
-		"ang":  "s",
-		"ie":   "d",
-		"ian":  "f",
-		"iang": "g",
-		"uang": "g",
-		"iong": "h",
-		"ong":  "h",
-		"er":   "j",
-		"iu":   "j",
-		"ei":   "k",
-		"uan":  "l",
-		"ing":  ";",
-		"ou":   "z",
-		"ia":   "x",
-		"ua":   "x",
-		"iao":  "b",
-		"ue":   "n",
-		"ui":   "n",
-		"un":   "m",
-	},
-}
+	doublePinyinSogou = schema{
+		name:            "cn_en_sogou",
+		desc:            "搜狗双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_sogou.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "oa",
+			"-e-":   "oe",
+			"-o-":   "oo",
+			"-ai-":  "ol",
+			"-ei-":  "oz",
+			"-ou-":  "ob",
+			"-an-":  "oj",
+			"-en-":  "of",
+			"-ang-": "oh",
+			"-eng-": "og",
+			"-ao-":  "ok",
+			"-er-":  "or",
+			// zh ch sh
+			"zh": "v",
+			"ch": "i",
+			"sh": "u",
+			// 韵母
+			"iu":   "q",
+			"ia":   "w",
+			"ua":   "w",
+			"er":   "r",
+			"uan":  "r",
+			"ue":   "t",
+			"uai":  "y",
+			"uo":   "o",
+			"un":   "p",
+			"iong": "s",
+			"ong":  "s",
+			"iang": "d",
+			"uang": "d",
+			"en":   "f",
+			"eng":  "g",
+			"ang":  "h",
+			"an":   "j",
+			"ao":   "k",
+			"ai":   "l",
+			"ing":  ";",
+			"ei":   "z",
+			"ie":   "x",
+			"iao":  "c",
+			"ui":   "v",
+			"ve":   "t",
+			"ou":   "b",
+			"in":   "n",
+			"ian":  "m",
+		},
+	}
 
-var doublePinyinABC = schema{
-	name:            "cn_en_abc",
-	desc:            "智能 ABC 双拼",
-	combinationType: "unique",
-	path:            filepath.Join(RimeDir, "en_dicts/cn_en_abc.txt"),
-	mapping: map[string]string{
-		// 零声母
-		"-a-":   "oa",
-		"-e-":   "oe",
-		"-o-":   "oo",
-		"-ai-":  "ol",
-		"-ei-":  "oq",
-		"-ou-":  "ob",
-		"-an-":  "oj",
-		"-en-":  "of",
-		"-ang-": "oh",
-		"-eng-": "og",
-		"-ao-":  "ok",
-		"-er-":  "or",
-		// zh ch sh
-		"zh": "a",
-		"ch": "e",
-		"sh": "v",
-		// 韵母
-		"ao":   "k",
-		"en":   "f",
-		"an":   "j",
-		"eng":  "g",
-		"in":   "c",
-		"uai":  "c",
-		"uo":   "o",
-		"ai":   "l",
-		"ang":  "h",
-		"ie":   "x",
-		"ian":  "w",
-		"iang": "t",
-		"uang": "t",
-		"iong": "s",
-		"ong":  "s",
-		"er":   "r",
-		"iu":   "r",
-		"ei":   "q",
-		"uan":  "p",
-		"ing":  "y",
-		"ou":   "b",
-		"ia":   "d",
-		"ua":   "d",
-		"iao":  "z",
-		"ue":   "m",
-		"ui":   "m",
-		"un":   "n",
-	},
+	doublePinyinZiGuang = schema{
+		name:            "cn_en_ziguang",
+		desc:            "紫光双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_ziguang.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "oa",
+			"-e-":   "oe",
+			"-o-":   "oo",
+			"-ai-":  "op",
+			"-ei-":  "ok",
+			"-ou-":  "oz",
+			"-an-":  "or",
+			"-en-":  "ow",
+			"-ang-": "os",
+			"-eng-": "ot",
+			"-ao-":  "oq",
+			"-er-":  "oj",
+			// zh ch sh
+			"zh": "u",
+			"ch": "a",
+			"sh": "i",
+			// 韵母
+			"ao":   "q",
+			"en":   "w",
+			"an":   "r",
+			"eng":  "t",
+			"in":   "y",
+			"uai":  "y",
+			"uo":   "o",
+			"ai":   "p",
+			"ang":  "s",
+			"ie":   "d",
+			"ian":  "f",
+			"iang": "g",
+			"uang": "g",
+			"iong": "h",
+			"ong":  "h",
+			"er":   "j",
+			"iu":   "j",
+			"ei":   "k",
+			"uan":  "l",
+			"ing":  ";",
+			"ou":   "z",
+			"ia":   "x",
+			"ua":   "x",
+			"iao":  "b",
+			"ue":   "n",
+			"ui":   "n",
+			"un":   "m",
+		},
+	}
+
+	doublePinyinABC = schema{
+		name:            "cn_en_abc",
+		desc:            "智能 ABC 双拼",
+		combinationType: "unique",
+		path:            filepath.Join(RimeDir, "en_dicts/cn_en_abc.txt"),
+		mapping: map[string]string{
+			// 零声母
+			"-a-":   "oa",
+			"-e-":   "oe",
+			"-o-":   "oo",
+			"-ai-":  "ol",
+			"-ei-":  "oq",
+			"-ou-":  "ob",
+			"-an-":  "oj",
+			"-en-":  "of",
+			"-ang-": "oh",
+			"-eng-": "og",
+			"-ao-":  "ok",
+			"-er-":  "or",
+			// zh ch sh
+			"zh": "a",
+			"ch": "e",
+			"sh": "v",
+			// 韵母
+			"ao":   "k",
+			"en":   "f",
+			"an":   "j",
+			"eng":  "g",
+			"in":   "c",
+			"uai":  "c",
+			"uo":   "o",
+			"ai":   "l",
+			"ang":  "h",
+			"ie":   "x",
+			"ian":  "w",
+			"iang": "t",
+			"uang": "t",
+			"iong": "s",
+			"ong":  "s",
+			"er":   "r",
+			"iu":   "r",
+			"ei":   "q",
+			"uan":  "p",
+			"ing":  "y",
+			"ou":   "b",
+			"ia":   "d",
+			"ua":   "d",
+			"iao":  "z",
+			"ue":   "m",
+			"ui":   "m",
+			"un":   "n",
+		},
+	}
 }
 
 // CnEn 从 others/cn_en.txt 生成全拼和各个双拼的中英混输词库
