@@ -236,9 +236,10 @@ var onlyOne = map[string]string{
 	"嘞":   "lei",
 	"凹":   "ao",
 	"给":   "gei",
+	"栖":   "qi",
 }
 
-func init() {
+func initPinyin() {
 	// 从 base、ext 准备结巴的词典和词组拼音映射
 	for _, dictPath := range []string{BasePath, ExtPath} {
 		file, err := os.Open(dictPath)
